requirements.txt
alembic==1.16.5
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.115.0
idna==3.10
lingua-language-detector==2.1.1
Mako==1.3.10
MarkupSafe==3.0.3
polib==1.2.0
pydantic==2.9.0
PyYAML==6.0.2
requests==2.32.5
soupsieve==2.8
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.32.0
